+++ pykechain/models/scope.py
@@ -1,97 +1,363 @@
-import datetime
-import warnings
-from typing import Any, Text, List  # noqa: F401
+from datetime import datetime
+from typing import Union, Text, Dict, Optional, List  # noqa: F401
 
 import requests
-from six import text_type, string_types
 
-from pykechain.enums import Multiplicity, ScopeStatus
-from pykechain.exceptions import APIError, NotFoundError, IllegalArgumentError, ForbiddenError
-from pykechain.models import Base
-from pykechain.utils import is_uuid, parse_datetime
+from pykechain.defaults import API_EXTRA_PARAMS
+from pykechain.enums import Multiplicity, ScopeStatus, SubprocessDisplayMode, KEChainPages, ScopeRoles, \
+    ScopeMemberActions
+from pykechain.exceptions import APIError, NotFoundError, IllegalArgumentError
+from pykechain.models.base import Base
+from pykechain.models.input_checks import check_text, check_datetime, check_enum, check_list_of_text, \
+    check_base, check_type
+from pykechain.models.representations.component import RepresentationsComponent
+from pykechain.models.sidebar.sidebar_manager import SideBarManager
+from pykechain.models.tags import TagsMixin
+from pykechain.models.team import Team
+from pykechain.utils import parse_datetime, find
 
 
-class Scope(Base):  # pragma: no cover
+class Scope(Base, TagsMixin):
     """A virtual object representing a KE-chain scope.
 
-    :ivar name: Name of the scope
-    :type name: str
-    :ivar id: UUID of the scope
-    :type if: uuid
-    :ivar workflow_root: Root of the workflow
+    :ivar id: id of the activity
+    :type id: uuid
+    :ivar name: name of the activity
+    :type name: basestring
+    :ivar created_at: created datetime of the activity
+    :type created_at: datetime
+    :ivar updated_at: updated datetime of the activity
+    :type updated_at: datetime
+    :ivar description: description of the activity
+    :type description: basestring
+    :ivar workflow_root: uuid of the workflow root object
     :type workflow_root: uuid
-    :ivar description: Description of the scope
-    :type description: str or None
-    :ivar status: Status of the scope, enumeration of `ScopeStatus`
-    :type status: str
-    :ivar start_date: Start date of the scope
-    :type start_date: datetime or None
-    :ivar due_date: Due date of the scope
-    :type due_date: datetime or None
-    :ivar type: Type of scope, enumeration of `ScopeType`
-    :type type: str
-    :ivar tags: list of unique tags of the scope.
-    :type tags: list of str
-    :ivar team: Team associated to the scope
-    :type team: `Team` or None
+    :ivar status: status of the scope. One of :class:`pykechain.enums.ScopeStatus`
+    :type status: basestring
+    :ivar type: Type of the Scope. One of :class:`pykechain.enums.ScopeType` for WIM version 2
+    :type type: basestring
     """
 
-    def __init__(self, json, **kwargs):
-        # type: (dict, **Any) -> None
+    def __init__(self, json: Dict, **kwargs) -> None:
         """Construct a scope from provided json data."""
-        super(Scope, self).__init__(json, **kwargs)
-
-        self.bucket = json.get('bucket', {})
+        super().__init__(json, **kwargs)
 
         # for 'kechain2.core.wim <2.0.0'
         self.process = json.get('process')
         # for 'kechain2.core.wim >=2.0.0'
         self.workflow_root = json.get('workflow_root_id')
 
-        self.description = json.get('description')
+        self._workflow_root_process = None
+        self._catalog_root_process = None
+        self._app_root_process = None
+        self._product_root_model = None
+        self._product_root_instance = None
+        self._catalog_root_model = None
+        self._catalog_root_instance = None
+
+        self.ref = json.get('ref')
+        self.description = json.get('text')
         self.status = json.get('status')
-        self.type = json.get('type')
+        self.category = json.get('category')
+
+        self._tags = json.get('tags')
 
         self.start_date = parse_datetime(json.get('start_date'))
         self.due_date = parse_datetime(json.get('due_date'))
 
-    def __repr__(self):  # pragma: no cover
-        return "<pyke Scope '{}' id {}>".format(self.name, self.id[-8:])
+        self._representations_container = RepresentationsComponent(
+            self,
+            self.options.get('representations', {}),
+            self._save_representations,
+        )
 
     @property
     def team(self):
         """Team to which the scope is assigned."""
-        team_dict = self._json_data.get('team')
+        team_dict = self._json_data.get('team_id_name')
         if team_dict and team_dict.get('id'):
             return self._client.team(pk=team_dict.get('id'))
         else:
             return None
 
     @property
-    def tags(self) -> List[Text]:
+    def options(self):
+        """Options of the Scope.
+
+        .. versionadded: 3.0
         """
-        Tags of the scope.
+        return self._json_data.get('scope_options')
 
-        A list of unique strings.
+    @options.setter
+    def options(self, option_value):
+        self.edit(options=option_value)
 
-        :return: list of tag strings
-        :rtype: list of str
+    def refresh(self, json=None, url=None, extra_params=None):
+        """Refresh the object in place."""
+        super().refresh(json=json,
+                        url=self._client._build_url('scope', scope_id=self.id),
+                        extra_params=API_EXTRA_PARAMS['scope'])
+
+    @property
+    def representations(self):
+        """Get and set the scope representations."""
+        return self._representations_container.get_representations()
+
+    @representations.setter
+    def representations(self, value):
+        self._representations_container.set_representations(value)
+
+    def _save_representations(self, representation_options):
+        options = self.options
+        options.update({'representations': representation_options})
+        self.options = options
+
+    @property
+    def workflow_root_process(self) -> 'Activity2':
+        """Retrieve the Activity2 root object with classification WORKFLOW."""
+        if self._workflow_root_process is None:
+            self._workflow_root_process = self.activity(id=self._json_data['workflow_root_id'])
+        return self._workflow_root_process
+
+    @property
+    def app_root_process(self) -> 'Activity2':
+        """Retrieve the Activity2 root object with classification APP."""
+        if self._app_root_process is None:
+            self._app_root_process = self.activity(id=self._json_data['app_root_id'])
+        return self._app_root_process
+
+    @property
+    def catalog_root_process(self) -> 'Activity2':
+        """Retrieve the Activity2 root object with classification CATALOG."""
+        if self._catalog_root_process is None:
+            self._catalog_root_process = self.activity(id=self._json_data['catalog_root_id'])
+        return self._catalog_root_process
+
+    @property
+    def product_root_model(self) -> 'Part2':
+        """Retrieve the Part2 root object with classification PRODUCT and category MODEL."""
+        if self._product_root_model is None:
+            self._product_root_model = self.model(id=self._json_data['product_model_id'])
+        return self._product_root_model
+
+    @property
+    def product_root_instance(self) -> 'Part2':
+        """Retrieve the Part2 root object with classification PRODUCT and category INSTANCE."""
+        if self._product_root_instance is None:
+            self._product_root_instance = self.part(id=self._json_data['product_instance_id'])
+        return self._product_root_instance
+
+    @property
+    def catalog_root_model(self) -> 'Part2':
+        """Retrieve the Part2 root object with classification CATALOG and category MODEL."""
+        if self._catalog_root_model is None:
+            self._catalog_root_model = self.model(id=self._json_data['catalog_model_id'])
+        return self._catalog_root_model
+
+    @property
+    def catalog_root_instance(self) -> 'Part2':
+        """Retrieve the Part2 root object with classification CATALOG and category INSTANCE."""
+        if self._catalog_root_instance is None:
+            self._catalog_root_instance = self.part(id=self._json_data['catalog_instance_id'])
+        return self._catalog_root_instance
+
+    #
+    # CRUD methods
+    #
+
+    def _update_scope_project_team(self, action, role, user):
         """
-        return self._json_data.get('tags')
+        Update the Project Team of the Scope. Updates include addition or removing of managers or members.
 
-    def parts(self, *args, **kwargs) -> 'PartSet':
+        :param action: type of action to be applied
+        :type action: ScopeMemberActions
+        :param role: type of role to be applied to the user
+        :type role: ScopeRoles
+        :param user: the username of the user to which the action applies to
+        :type user: basestring
+        :raises APIError: When unable to update the scope project team.
+        """
+        action = check_enum(action, ScopeMemberActions, 'action')
+        role = check_enum(role, ScopeRoles, 'role')
+        user = check_text(user, 'user')
+
+        users = self._client._retrieve_users()['results']  # type: List[Dict]
+        user_object = find(users, lambda u: u['username'] == user)  # type: Dict
+        if user_object is None:
+            raise NotFoundError('User "{}" does not exist'.format(user))
+
+        url = self._client._build_url('scope_{}_{}'.format(action, role), scope_id=self.id)
+
+        response = self._client._request('PUT', url,
+                                         params=API_EXTRA_PARAMS[self.__class__.__name__.lower()],
+                                         data={'user_id': user_object['pk']})
+
+        if response.status_code != requests.codes.ok:  # pragma: no cover
+            raise APIError("Could not {} {} in Scope".format(action, role), response=response)
+
+        self.refresh(json=response.json().get('results')[0])
+
+    def edit(
+            self,
+            name: Optional[Text] = None,
+            description: Optional[Text] = None,
+            start_date: Optional[datetime] = None,
+            due_date: Optional[datetime] = None,
+            status: Optional[Union[Text, ScopeStatus]] = None,
+            tags: Optional[List[Text]] = None,
+            team: Optional[Union[Team, Text]] = None,
+            options: Optional[Dict] = None,
+    ) -> None:
+        """Edit the details of a scope.
+
+        :param name: (optionally) edit the name of the scope
+        :type name: basestring or None
+        :param description: (optionally) edit the description of the scope
+        :type description: basestring or None
+        :param start_date: (optionally) edit the start date of the scope as a datetime object (UTC time/timezone
+                            aware preferred)
+        :type start_date: datetime or None
+        :param due_date: (optionally) edit the due_date of the scope as a datetime object (UTC time/timzeone
+                            aware preferred)
+        :type due_date: datetime or None
+        :param status: (optionally) edit the status of the scope as a string based
+        :type status: basestring or None
+        :param tags: (optionally) replace the tags on a scope, which is a list of strings ["one","two","three"]
+        :type tags: list of basestring or None
+        :param team: (optionally) add the scope to a team
+        :type team: UUIDstring or None
+        :param options: (optionally) custom options dictionary stored on the scope object
+        :type options: dict or None
+        :raises IllegalArgumentError: if the type of the inputs is not correct
+        :raises APIError: if another Error occurs
+        :warns: UserWarning - When a naive datetime is provided. Defaults to UTC.
+
+        Examples
+        --------
+        >>> from datetime import datetime
+        >>> project.edit(name='New project name',
+        ...              description='Changing the description just because I can',
+        ...              start_date=datetime.now(),  # naive time is interpreted as UTC time
+        ...              status=ScopeStatus.CLOSED)
+
+        If we want to provide timezone aware datetime objects we can use the 3rd party convenience library :mod:`pytz`.
+        Mind that we need to fetch the timezone first and use `<timezone>.localize(<your datetime>)` to make it
+        work correctly.
+
+        Using `datetime(2017,6,1,23,59,0 tzinfo=<tz>)` does NOT work for most timezones with a
+        daylight saving time. Check the `pytz <http://pythonhosted.org/pytz/#localized-times-and-date-arithmetic>`_
+        documentation.
+
+        To make it work using :mod:`pytz` and timezone aware :mod:`datetime` see the following example::
+
+        >>> import pytz
+        >>> start_date_tzaware = datetime.now(pytz.utc)
+        >>> mytimezone = pytz.timezone('Europe/Amsterdam')
+        >>> due_date_tzaware = mytimezone.localize(datetime(2019, 10, 27, 23, 59, 0))
+        >>> project.edit(due_date=due_date_tzaware, start_date=start_date_tzaware)
+
+        To assign a scope to a team see the following example::
+
+        >>> my_team = client.team(name='My own team')
+        >>> project.edit(team=my_team)
+
+        """
+        update_dict = {
+            'id': self.id,
+            'name': check_text(name, 'name') or self.name,
+            'text': check_text(description, 'description') or self.description,
+            'start_date': check_datetime(start_date, 'start_date'),
+            'due_date': check_datetime(due_date, 'due_date'),
+            'status': check_enum(status, ScopeStatus, 'status') or self.status,
+        }
+        team = check_base(team, Team, 'team', method=self._client.team)
+        if team:
+            update_dict['team_id'] = team
+        tags = check_list_of_text(tags, 'tags', True)
+        if tags is not None:
+            update_dict['tags'] = tags
+        scope_options = check_type(options, dict, 'options')
+        if scope_options:
+            update_dict['scope_options'] = scope_options
+
+        url = self._client._build_url('scope', scope_id=self.id)
+
+        response = self._client._request('PUT', url,
+                                         params=API_EXTRA_PARAMS[self.__class__.__name__.lower()],
+                                         json=update_dict)
+
+        if response.status_code != requests.codes.ok:  # pragma: no cover
+            raise APIError("Could not update Scope {}".format(self), response=response)
+
+        self.refresh(json=response.json().get('results')[0])
+
+        # TODO tags that are set are not in response
+        if tags is not None:
+            self._tags = tags
+
+    def clone(self, **kwargs):
+        """Clone a scope.
+
+        See :method:`pykechain.Client.clone_scope()` for available parameters.
+        """
+        return self._client.clone_scope(source_scope=self, **kwargs)
+
+    def delete(self, asynchronous=True):
+        """Delete the scope.
+
+        Only works with enough permissions.
+
+        .. versionadded: 3.0
+
+        See :method:`pykechain.Client.delete_scope()` for available parameters.
+        :raises ForbiddenError: if you do not have the permissions to delete a scope
+        """
+        return self._client.delete_scope(scope=self, asynchronous=asynchronous)
+
+    #
+    # Part methods
+    #
+
+    def parts(self, *args, **kwargs):
         """Retrieve parts belonging to this scope.
+
+        This uses
 
         See :class:`pykechain.Client.parts` for available parameters.
         """
-        return self._client.parts(*args, bucket=self.bucket.get('id'), **kwargs)
+        return self._client.parts(*args, scope_id=self.id, **kwargs)
 
-    def part(self, *args, **kwargs) -> 'Part2':
+    def part(self, *args, **kwargs):
         """Retrieve a single part belonging to this scope.
 
         See :class:`pykechain.Client.part` for available parameters.
         """
-        return self._client.part(*args, bucket=self.bucket.get('id'), **kwargs)
+        return self._client.part(*args, scope_id=self.id, **kwargs)
+
+    def properties(self, *args, **kwargs):
+        """Retrieve properties belonging to this scope.
+
+        .. versionadded: 3.0
+
+        See :class:`pykechain.Client.properties` for available parameters.
+        """
+        return self._client.properties(*args, scope_id=self.id, **kwargs)
+
+    def property(self, *args, **kwargs):
+        """Retrieve a single property belonging to this scope.
+
+        .. versionadded: 3.0
+
+        See :class:`pykechain.Client.property` for available parameters.
+        """
+        return self._client.property(*args, scope_id=self.id, **kwargs)
+
+    def model(self, *args, **kwargs):
+        """Retrieve a single model belonging to this scope.
+
+        See :class:`pykechain.Client.model` for available parameters.
+        """
+        return self._client.model(*args, scope_id=self.id, **kwargs)
 
     def create_model(self, parent, name, multiplicity=Multiplicity.ZERO_MANY):
         """Create a single part model in this scope.
@@ -100,42 +366,78 @@ class Scope(Base):  # pragma: no cover
         """
         return self._client.create_model(parent, name, multiplicity=multiplicity)
 
-    def model(self, *args, **kwargs):
-        """Retrieve a single model belonging to this scope.
+    def create_model_with_properties(self, parent, name, multiplicity=Multiplicity.ZERO_MANY,
+                                     properties_fvalues=None, **kwargs):
+        """Create a model with its properties in a single API request.
 
-        See :class:`pykechain.Client.model` for available parameters.
+        See :func:`pykechain.Client.create_model_with_properties()` for available parameters.
         """
-        return self._client.model(*args, bucket=self.bucket.get('id'), **kwargs)
+        return self._client.create_model_with_properties(parent, name, multiplicity=multiplicity,
+                                                         properties_fvalues=properties_fvalues, **kwargs)
+
+    #
+    # Activity methods
+    #
 
     def activities(self, *args, **kwargs):
         """Retrieve activities belonging to this scope.
 
         See :class:`pykechain.Client.activities` for available parameters.
         """
-        if self._client.match_app_version(label='wim', version='<2.0.0', default=True):
-            return self._client.activities(*args, scope=self.id, **kwargs)
-        else:
-            return self._client.activities(*args, scope_id=self.id, **kwargs)
+        return self._client.activities(*args, scope_id=self.id, **kwargs)
 
     def activity(self, *args, **kwargs):
         """Retrieve a single activity belonging to this scope.
 
         See :class:`pykechain.Client.activity` for available parameters.
         """
-        if self._client.match_app_version(label='wim', version='<2.0.0', default=True):
-            return self._client.activity(*args, scope=self.id, **kwargs)
-        else:
-            return self._client.activity(*args, scope_id=self.id, **kwargs)
+        return self._client.activity(*args, scope_id=self.id, **kwargs)
 
     def create_activity(self, *args, **kwargs):
         """Create a new activity belonging to this scope.
 
         See :class:`pykechain.Client.create_activity` for available parameters.
         """
-        if self._client.match_app_version(label='wim', version='<2.0.0', default=True):
-            return self._client.create_activity(self.process, *args, **kwargs)
+        return self._client.create_activity(self.workflow_root, *args, **kwargs)
+
+    def side_bar(self, *args, **kwargs) -> Optional[SideBarManager]:
+        """Retrieve the side-bar manager."""
+        return SideBarManager(scope=self, *args, **kwargs)
+
+    def set_landing_page(self,
+                         activity: Union['Activity2', KEChainPages],
+                         task_display_mode: Optional[SubprocessDisplayMode] = SubprocessDisplayMode.ACTIVITIES,
+                         ) -> None:
+        """
+        Update the landing page of the scope.
+
+        :param activity: Activity2 object or KEChainPages option
+        :type activity: (Activity2, KEChainPages)
+        :param task_display_mode: display mode of the activity in KE-chain
+        :type task_display_mode: SubprocessDisplayMode
+        :return: None
+        :rtype None
+        """
+        from pykechain.models import Activity2
+
+        if not (isinstance(activity, Activity2) or activity in KEChainPages.values()):
+            raise IllegalArgumentError(
+                'activity must be of class Activity2 or a KEChainPages option, "{}" is not.'.format(activity))
+
+        check_enum(task_display_mode, SubprocessDisplayMode, 'task_display_mode')
+
+        if isinstance(activity, Activity2):
+            url = '#/scopes/{}/{}/{}'.format(self.id, task_display_mode, activity.id)
         else:
-            return self._client.create_activity(self.workflow_root, *args, **kwargs)
+            url = '#/scopes/{}/{}'.format(self.id, activity)
+
+        options = dict(self.options)
+        options.update({'landingPage': url})
+        self.options = options
+
+    #
+    # Service Methods
+    #
 
     def services(self, *args, **kwargs):
         """Retrieve services belonging to this scope.
@@ -182,27 +484,37 @@ class Scope(Base):  # pragma: no cover
         """
         return self._client.service_execution(*args, scope=self.id, **kwargs)
 
-    def members(self, is_manager=None):
+    #
+    # User and Members of the Scope
+    #
+
+    def members(self, is_manager: Optional[bool] = None, is_leadmember: Optional[bool] = None) -> List[Dict]:
         """
         Retrieve members of the scope.
 
-        :param is_manager: (optional) set to True to return only Scope members that are also managers.
+        :param is_manager: (otional) set to True/False to filter members that are/aren't managers, resp.
         :type is_manager: bool
-        :return: List of members (usernames)
+        :param is_leadmember: (optional) set to True/False to filter members that are/aren't leadmembers, resp.
+        :type is_leadmember: bool
+        :return: List of members, each defined as a dict
 
         Examples
         --------
         >>> members = project.members()
         >>> managers = project.members(is_manager=True)
+        >>> leadmembers = project.members(is_leadmember=True)
 
         """
-        if not is_manager:
-            return [member for member in self._json_data['members'] if member['is_active']]
-        else:
-            return [member for member in self._json_data['members'] if
-                    member.get('is_active', False) and member.get('is_manager', False)]
+        members = [member for member in self._json_data['members'] if member['is_active']]
 
-    def add_member(self, member):
+        if is_manager is not None:
+            members = [member for member in members if member.get('is_manager') == is_manager]
+        if is_leadmember is not None:
+            members = [member for member in members if member.get('is_leadmember') == is_leadmember]
+
+        return members
+
+    def add_member(self, member: Text) -> None:
         """
         Add a single member to the scope.
 
@@ -212,11 +524,9 @@ class Scope(Base):  # pragma: no cover
         :type member: basestring
         :raises APIError: when unable to update the scope member
         """
-        select_action = 'add_member'
+        self._update_scope_project_team(action=ScopeMemberActions.ADD, role=ScopeRoles.MEMBER, user=member)
 
-        self._update_scope_project_team(select_action=select_action, user=member, user_type='member')
-
-    def remove_member(self, member):
+    def remove_member(self, member: Text) -> None:
         """
         Remove a single member to the scope.
 
@@ -224,11 +534,9 @@ class Scope(Base):  # pragma: no cover
         :type member: basestring
         :raises APIError: when unable to update the scope member
         """
-        select_action = 'remove_member'
+        self._update_scope_project_team(action=ScopeMemberActions.REMOVE, role=ScopeRoles.MEMBER, user=member)
 
-        self._update_scope_project_team(select_action=select_action, user=member, user_type='member')
-
-    def add_manager(self, manager):
+    def add_manager(self, manager: Text) -> None:
         """
         Add a single manager to the scope.
 
@@ -236,11 +544,9 @@ class Scope(Base):  # pragma: no cover
         :type manager: basestring
         :raises APIError: when unable to update the scope manager
         """
-        select_action = 'add_manager'
+        self._update_scope_project_team(action=ScopeMemberActions.ADD, role=ScopeRoles.MANAGER, user=manager)
 
-        self._update_scope_project_team(select_action=select_action, user=manager, user_type='manager')
-
-    def remove_manager(self, manager):
+    def remove_manager(self, manager: Text) -> None:
         """
         Remove a single manager to the scope.
 
@@ -248,190 +554,24 @@ class Scope(Base):  # pragma: no cover
         :type manager: basestring
         :raises APIError: when unable to update the scope manager
         """
-        select_action = 'remove_manager'
+        self._update_scope_project_team(action=ScopeMemberActions.REMOVE, role=ScopeRoles.MANAGER, user=manager)
 
-        self._update_scope_project_team(select_action=select_action, user=manager, user_type='manager')
-
-    def _update_scope_project_team(self, select_action, user, user_type):
+    def add_leadmember(self, leadmember: Text) -> None:
         """
-        Update the Project Team of the Scope. Updates include addition or removing of managers or members.
+        Add a single leadmember to the scope.
 
-        :param select_action: type of action to be applied
-        :type select_action: basestring
-        :param user: the username of the user to which the action applies to
-        :type user: basestring
-        :param user_type: the type of the user (member or manager)
-        :type user_type: basestring
-        :raises APIError: When unable to update the scope project team.
+        :param leadmember: single username to be added to the scope list of leadmembers
+        :type leadmember: basestring
+        :raises APIError: when unable to update the scope leadmember
         """
-        if isinstance(user, (string_types, text_type)):
-            users = self._client._retrieve_users()
-            manager_object = next((item for item in users['results'] if item["username"] == user), None)
-            if manager_object:
-                url = self._client._build_url('scope', scope_id=self.id)
-                response = self._client._request('PUT', url, params={'select_action': select_action},
-                                                 data={'user_id': manager_object['pk']})
-                if response.status_code != requests.codes.ok:  # pragma: no cover
-                    raise APIError("Could not {} {} in Scope".format(select_action.split('_')[0], user_type))
-            else:
-                raise NotFoundError("User {} does not exist".format(user))
-        else:
-            raise TypeError("User {} should be defined as a string".format(user))
+        self._update_scope_project_team(action=ScopeMemberActions.ADD, role=ScopeRoles.LEADMEMBER, user=leadmember)
 
-    def _edit(self, update_dict):
-        url = self._client._build_url('scope', scope_id=self.id)
-
-        response = self._client._request('PUT', url, json=update_dict)
-
-        if response.status_code != requests.codes.ok:  # pragma: no cover
-            raise APIError("Could not update Scope ({})".format(response))
-        else:
-            self._json_data = response.json().get('results') and response.json().get('results')[0]
-
-    def edit(self, name=None, description=None, start_date=None, due_date=None, status=None, tags=None, team=None,
-             options=None, **kwargs):
-        """Edit the details of a scope.
-
-        :param name: (optionally) edit the name of the scope
-        :type name: basestring or None
-        :param description: (optionally) edit the description of the scope
-        :type description: basestring or None
-        :param start_date: (optionally) edit the start date of the scope as a datetime object (UTC time/timezone
-                            aware preferred)
-        :type start_date: datetime or None
-        :param due_date: (optionally) edit the due_date of the scope as a datetime object (UTC time/timzeone
-                            aware preferred)
-        :type due_date: datetime or None
-        :param status: (optionally) edit the status of the scope as a string based
-        :type status: basestring or None
-        :param tags: (optionally) replace the tags on a scope, which is a list of strings ["one","two","three"]
-        :type tags: list of basestring or None
-        :param team: (optionally) add the scope to a team
-        :type team: UUIDstring or None
-        :param options: (optionally) custom options dictionary stored on the scope object
-        :type options: dict or None
-        :raises IllegalArgumentError: if the type of the inputs is not correct
-        :raises APIError: if another Error occurs
-        :warns: UserWarning - When a naive datetime is provided. Defaults to UTC.
-
-        Examples
-        --------
-        >>> from datetime import datetime
-        >>> project.edit(name='New project name',
-        ...              description='Changing the description just because I can',
-        ...              start_date=datetime.utcnow(),  # naive time is interpreted as UTC time
-        ...              status=ScopeStatus.CLOSED)
-
-        If we want to provide timezone aware datetime objects we can use the 3rd party convenience library :mod:`pytz`.
-        Mind that we need to fetch the timezone first and use `<timezone>.localize(<your datetime>)` to make it
-        work correctly.
-
-        Using `datetime(2017,6,1,23,59,0 tzinfo=<tz>)` does NOT work for most timezones with a
-        daylight saving time. Check the `pytz <http://pythonhosted.org/pytz/#localized-times-and-date-arithmetic>`_
-        documentation.
-
-        To make it work using :mod:`pytz` and timezone aware :mod:`datetime` see the following example::
-
-        >>> import pytz
-        >>> start_date_tzaware = datetime.now(pytz.utc)
-        >>> mytimezone = pytz.timezone('Europe/Amsterdam')
-        >>> due_date_tzaware = mytimezone.localize(datetime(2019, 10, 27, 23, 59, 0))
-        >>> project.edit(due_date=due_date_tzaware, start_date=start_date_tzaware)
-
-        To assign a scope to a team see the following example::
-
-        >>> my_team = client.team(name='My own team')
-        >>> project.edit(team=my_team)
-
+    def remove_leadmember(self, leadmember: Text) -> None:
         """
-        update_dict = {'id': self.id}
-        if name is not None:
-            if isinstance(name, (str, text_type)):
-                update_dict.update({'name': name})
-                self.name = name
-            else:
-                raise IllegalArgumentError('Name should be a string')
+        Remove a single leadmember to the scope.
 
-        if description is not None:  # isinstance(description, (str, text_type)):
-            if isinstance(description, (str, text_type)):
-                update_dict.update({'text': description})
-                self.text = description
-            else:
-                raise IllegalArgumentError('Description should be a string')
-
-        if start_date is not None:
-            if isinstance(start_date, datetime.datetime):
-                if not start_date.tzinfo:
-                    warnings.warn("The startdate '{}' is naive and not timezone aware, use pytz.timezone info. "
-                                  "This date is interpreted as UTC time.".format(start_date.isoformat(sep=' ')))
-                update_dict.update({'start_date': start_date.isoformat(sep='T')})
-            else:
-                raise IllegalArgumentError('Start date should be a datetime.datetime() object')
-
-        if due_date is not None:
-            if isinstance(due_date, datetime.datetime):
-                if not due_date.tzinfo:
-                    warnings.warn("The duedate '{}' is naive and not timezone aware, use pytz.timezone info. "
-                                  "This date is interpreted as UTC time.".format(due_date.isoformat(sep=' ')))
-                update_dict.update({'due_date': due_date.isoformat(sep='T')})
-            else:
-                raise IllegalArgumentError('Due date should be a datetime.datetime() object')
-
-        if status is not None:
-            if isinstance(status, (str, text_type)) and status in ScopeStatus.values():
-                update_dict.update({'status': status})
-            else:
-                raise IllegalArgumentError('Status should be a string and in the list of acceptable '
-                                           'status strings: {}'.format(ScopeStatus.values()))
-
-        if tags is not None:
-            if isinstance(tags, (list, tuple, set)):
-                update_dict.update({'tags': tags})
-            else:
-                raise IllegalArgumentError('tags should be a an array (list, tuple, set) of strings')
-
-        if team is not None:
-            from pykechain.models import Team
-            if isinstance(team, (str, text_type)) and is_uuid(team):
-                update_dict.update({'team_id': team})
-            elif isinstance(team, Team):
-                update_dict.update({'team_id': team.id})
-            else:
-                raise IllegalArgumentError("team should be the uuid of a team")
-
-        if options is not None:
-            if isinstance(options, dict):
-                update_dict.update({'options': options})
-            else:
-                raise IllegalArgumentError("options should be a dictionary")
-
-        # do the update itself in an abstracted function.
-        self._edit(update_dict)
-
-    def clone(self, *args, **kwargs):
+        :param leadmember: single username to be added to the scope list of leadmembers
+        :type leadmember: basestring
+        :raises APIError: when unable to update the scope leadmember
         """
-        Clone current scope.
-
-        See :class:`pykechain.Client.clone_scope` for available parameters.
-
-        .. versionadded:: 2.6.0
-        """
-        return self._client.clone_scope(*args, source_scope=self, **kwargs)
-
-    def delete(self):
-        """
-        Delete the scope.
-
-        Only works with enough permissions.
-
-        .. versionadded: 3.0
-
-        :raises ForbiddenError: if you do not have the permissions to delete a scope
-        """
-        url = self._client._build_url('scope', scope_id=self.id)
-        response = self._client._request('DELETE', url)
-
-        if response.status_code != requests.codes.no_content:  # pragma: no cover
-            if response.status_code == requests.codes.forbiddem:
-                raise ForbiddenError("Forbidden to delete scope, {}: {}".format(str(response), response.content))
-            raise APIError("Could not delete scope, {}: {}".format(str(response), response.content))
+        self._update_scope_project_team(action=ScopeMemberActions.REMOVE, role=ScopeRoles.LEADMEMBER, user=leadmember)
